+++ src/Repository/PersonaRepository.php
@@ -6,7 +6,7 @@ use Phamily\Framework\Model\PersonaInterface;
 use Zend\Db\TableGateway\TableGateway;
 use Phamily\Framework\Model\Persona;
 use Phamily\Framework\Repository\exceptions\NotFoundException;
-use Phamily\Framework\traits\BitmaskTrait;
+use Phamily\Framework\Util\BitmaskTrait;
 use Phamily\Framework\Repository\conditions\SiblingsQueryCondition;
 use Zend\Db\TableGateway\Feature\SequenceFeature;