+++ src/Util/BitmaskTrait.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Phamily\Framework\Util;
+
+/**
+ *
+ * @author samizdam
+ *
+ */
+trait BitmaskTrait
+{
+
+    /**
+     *
+     * @param int $options
+     * @param int $flag
+     * @return boolean
+     */
+    protected function isFlagSet($options, $flag)
+    {
+        return (($options & $flag) == $flag);
+    }
+}